moved attacking squares forwarding to chess-board.component
